moved sia/polyOperation apis
