+++ frontend/lost/src/containers/Datasets/ReviewPage.tsx
@@ -2,7 +2,7 @@ import { useParams } from 'react-router-dom'
 import SiaWrapper from '../Annotation/SIA/SiaWrapper'
 import { CSSProperties, useEffect, useState } from 'react'
 import siaApi, { ReviewData, useReview } from '../../api/dataset_review'
-import { ImageSwitchData } from '../../actions/sia/sia_api'
+import { ImageSwitchData } from '../../api/sia'
 import { useAnnotask } from '../../api/anno_task'
 import AnnotationTop from '../Annotation/AnnoTask/AnnotationTop'
